Reorganize glut (preparation for glfw)
