+++ dart/gui/CMakeLists.txt
@@ -21,15 +21,31 @@ else()
 endif()
 
 # Search all header and source files
-file(GLOB hdrs "*.hpp")
-file(GLOB srcs "*.cpp")
+file(GLOB hdrs "*.hpp" "*.h" "detail/*.hpp")
+file(GLOB srcs "*.cpp" "*.c" "detail/*.cpp")
+
+function(dart_add_gui_headers)
+  dart_property_add(DART_GUI_HEADERS ${ARGN})
+endfunction()
+
+function(dart_add_gui_sources)
+  dart_property_add(DART_GUI_SOURCES ${ARGN})
+endfunction()
+
+# Add required subdirectory
+add_subdirectory(glut)
+
+get_property(dart_gui_headers GLOBAL PROPERTY DART_GUI_HEADERS)
+get_property(dart_gui_sources GLOBAL PROPERTY DART_GUI_SOURCES)
 
 # Set local target name
 set(target_name ${PROJECT_NAME}-gui)
 set(component_name gui)
 
 # Add target
-dart_add_library(${target_name} ${hdrs} ${srcs})
+dart_add_library(${target_name}
+  ${hdrs} ${srcs} ${dart_gui_headers} ${dart_gui_sources}
+)
 target_include_directories(
   ${target_name} SYSTEM
   PUBLIC
@@ -62,6 +78,11 @@ add_subdirectory(osg)
 
 # Generate header for this namespace
 dart_get_filename_components(header_names "gui headers" ${hdrs})
+set(
+  header_names
+  ${header_names}
+  glut/glut.hpp
+)
 dart_generate_include_header_list(
   gui_headers
   "dart/gui/"
