+++ dart/gui/GLFuncs.hpp
@@ -34,15 +34,18 @@
 #define DART_GUI_GLFUNCS_HPP_
 
 #include <string>
-
 #include <Eigen/Eigen>
+#include "dart/common/Deprecated.hpp"
 
 namespace dart {
 namespace gui {
 
-/// \brief
-  void drawStringOnScreen(float _x, float _y, const std::string& _s,
-                          bool _bigFont = true);
+/// \deprecated Deprecated in 6.1. Please use
+/// dart::gui::glut::drawStringOnScreen() instead in
+/// dart/gui/glut/GLUTFuncs.hpp file.
+DART_DEPRECATED(6.1)
+void drawStringOnScreen(float _x, float _y, const std::string& _s,
+                        bool _bigFont = true);
 
 /// \brief
 void drawArrow3D(const Eigen::Vector3d& _pt, const Eigen::Vector3d& _dir,
