+++ dart/gui/SimWindow.hpp
@@ -39,120 +39,15 @@
 #ifndef DART_GUI_SIMWINDOW_HPP_
 #define DART_GUI_SIMWINDOW_HPP_
 
-#include <vector>
+#warning "This file is deprecated in DART 6.1. "\
+         "Please use dart/gui/glut/SimWindow.hpp instead."
 
-#include <Eigen/Dense>
-
-#include "dart/common/Deprecated.hpp"
-#include "dart/gui/Win3D.hpp"
-#include "dart/simulation/World.hpp"
+#include "dart/gui/glut/SimWindow.hpp"
 
 namespace dart {
 namespace gui {
 
-class GraphWindow;
-
-/// \brief
-class SimWindow : public Win3D {
-public:
-  /// \brief
-  SimWindow();
-
-  /// \brief
-  virtual ~SimWindow();
-
-  /// \brief
-  virtual void timeStepping();
-
-  virtual void drawWorld() const;
-
-  virtual void drawSkeletons() const;
-
-  DART_DEPRECATED(6.0)
-  virtual void drawSkels();
-
-  DART_DEPRECATED(6.0)
-  virtual void drawEntities();
-
-  /// \brief
-  void displayTimer(int _val) override;
-
-  /// \brief
-  void draw() override;
-
-  /// \brief
-  void keyboard(unsigned char _key, int _x, int _y) override;
-
-  /// \brief
-  void setWorld(dart::simulation::WorldPtr _world);
-
-  /// \brief Save world in 'tempWorld.txt'
-  void saveWorld();
-
-  /// \brief Plot _data in a 2D window
-  void plot(Eigen::VectorXd& _data);
-//  bool isSimulating() const { return mSimulating; }
-
-//  void setSimulatingFlag(int _flag) { mSimulating = _flag; }
-
-protected:
-
-  virtual void drawSkeleton(
-      const dynamics::Skeleton* skeleton,
-      const Eigen::Vector4d& color = Eigen::Vector4d::Constant(0.5),
-      bool useDefaultColor = true) const;
-
-  virtual void drawEntity(
-      const dynamics::Entity* entity,
-      const Eigen::Vector4d& color = Eigen::Vector4d::Constant(0.5),
-      bool useDefaultColor = true) const;
-
-  virtual void drawBodyNode(
-      const dynamics::BodyNode* bodyNode,
-      const Eigen::Vector4d& color = Eigen::Vector4d::Constant(0.5),
-      bool useDefaultColor = true,
-      bool recursive = false) const;
-
-  virtual void drawShapeFrame(
-      const dynamics::ShapeFrame* shapeFrame,
-      const Eigen::Vector4d& color = Eigen::Vector4d::Constant(0.5),
-      bool useDefaultColor = true) const;
-
-  virtual void drawShape(
-      const dynamics::Shape* shape,
-      const Eigen::Vector4d& color = Eigen::Vector4d::Constant(0.5)) const;
-
-  virtual void drawPointMasses(
-      const std::vector<dynamics::PointMass*> pointMasses,
-      const Eigen::Vector4d& color = Eigen::Vector4d::Constant(0.5),
-      bool useDefaultColor = true) const;
-
-  virtual void drawMarker(
-      const dynamics::Marker* marker,
-      const Eigen::Vector4d& color = Eigen::Vector4d::Constant(0.5),
-      bool useDefaultColor = true) const;
-
-  /// \brief
-  simulation::WorldPtr mWorld;
-
-  /// \brief
-  int mPlayFrame;
-
-  /// \brief
-  bool mPlay;
-
-  /// \brief
-  bool mSimulating;
-
-  /// If true, render point masses of soft bodies
-  bool mShowPointMasses;
-
-  /// If true, render markers
-  bool mShowMarkers;
-
-  /// \brief Array of graph windows
-  std::vector<GraphWindow*> mGraphWindows;
-};
+using SimWindow = ::dart::gui::glut::SimWindow;
 
 }  // namespace gui
 }  // namespace dart
