+++ dart/gui/Win3D.hpp
@@ -33,43 +33,15 @@
 #ifndef DART_GUI_WIN3D_HPP_
 #define DART_GUI_WIN3D_HPP_
 
-#include <Eigen/Eigen>
+#warning "This file is deprecated in DART 6.1. "\
+         "Please use dart/gui/glut/Win3D.hpp instead."
 
-#include "dart/gui/GlutWindow.hpp"
-#include "dart/gui/Trackball.hpp"
+#include "dart/gui/glut/Win3D.hpp"
 
 namespace dart {
 namespace gui {
 
-class Win3D : public GlutWindow {
-public:
-  Win3D();
-
-  void initWindow(int _w, int _h, const char* _name) override;
-  void resize(int _w, int _h) override;
-  void render() override;
-
-  void keyboard(unsigned char _key, int _x, int _y) override;
-  void click(int _button, int _state, int _x, int _y) override;
-  void drag(int _x, int _y) override;
-
-  virtual void initGL();
-  virtual void initLights();
-
-  virtual void draw()=0;
-
-protected:
-  Trackball mTrackBall;
-  Eigen::Vector3d mTrans;
-  Eigen::Vector3d mEye;
-  Eigen::Vector3d mUp;
-  float mZoom;
-  float mPersp;
-
-  bool mRotate;
-  bool mTranslate;
-  bool mZooming;
-};
+using Win3D = ::dart::gui::glut::Win3D;
 
 }  // namespace gui
 }  // namespace dart
