+++ dart/gui/glut/GlutWindow.cpp
@@ -30,7 +30,7 @@
  *   POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "dart/gui/GlutWindow.hpp"
+#include "dart/gui/glut/GlutWindow.hpp"
 
 #include "dart/external/lodepng/lodepng.h"
 
@@ -48,17 +48,18 @@
 #include <vector>
 
 #include "dart/common/Console.hpp"
-#include "dart/gui/LoadGlut.hpp"
+#include "dart/gui/glut/LoadGlut.hpp"
 #include "dart/gui/GLFuncs.hpp"
 #include "dart/gui/OpenGLRenderInterface.hpp"
 
 namespace dart {
 namespace gui {
+namespace glut {
 
-std::vector<GlutWindow*> GlutWindow::mWindows;
-std::vector<int> GlutWindow::mWinIDs;
+std::vector<Window*> Window::mWindows;
+std::vector<int> Window::mWinIDs;
 
-GlutWindow::GlutWindow() {
+Window::Window() {
   mWinWidth = 0;
   mWinHeight = 0;
   mMouseX = 0;
@@ -74,11 +75,11 @@ GlutWindow::GlutWindow() {
   mRI = nullptr;
 }
 
-GlutWindow::~GlutWindow() {
+Window::~Window() {
   delete mRI;
 }
 
-void GlutWindow::initWindow(int _w, int _h, const char* _name) {
+void Window::initWindow(int _w, int _h, const char* _name) {
   mWindows.push_back(this);
 
   mWinWidth = _w;
@@ -113,53 +114,53 @@ void GlutWindow::initWindow(int _w, int _h, const char* _name) {
   // Note: We book the timer id 0 for the main rendering purpose.
 }
 
-void GlutWindow::reshape(int _w, int _h) {
+void Window::reshape(int _w, int _h) {
   current()->mScreenshotTemp = std::vector<unsigned char>(_w*_h*4);
   current()->mScreenshotTemp2 = std::vector<unsigned char>(_w*_h*4);
   current()->resize(_w, _h);
 }
 
-void GlutWindow::keyEvent(unsigned char _key, int _x, int _y) {
+void Window::keyEvent(unsigned char _key, int _x, int _y) {
   current()->keyboard(_key, _x, _y);
 }
 
-void GlutWindow::specKeyEvent(int _key, int _x, int _y) {
+void Window::specKeyEvent(int _key, int _x, int _y) {
   current()->specKey(_key, _x, _y);
 }
 
-void GlutWindow::mouseClick(int _button, int _state, int _x, int _y) {
+void Window::mouseClick(int _button, int _state, int _x, int _y) {
   current()->click(_button, _state, _x, _y);
 }
 
-void GlutWindow::mouseDrag(int _x, int _y) {
+void Window::mouseDrag(int _x, int _y) {
   current()->drag(_x, _y);
 }
 
-void GlutWindow::mouseMove(int _x, int _y) {
+void Window::mouseMove(int _x, int _y) {
   current()->move(_x, _y);
 }
 
-void GlutWindow::refresh() {
+void Window::refresh() {
   current()->render();
 }
 
-void GlutWindow::refreshTimer(int _val) {
+void Window::refreshTimer(int _val) {
   current()->displayTimer(_val);
 }
 
-void GlutWindow::displayTimer(int _val) {
+void Window::displayTimer(int _val) {
   glutPostRedisplay();
   glutTimerFunc(mDisplayTimeout, refreshTimer, _val);
 }
 
-void GlutWindow::simTimer(int /*_val*/) {
+void Window::simTimer(int /*_val*/) {
 }
 
-void GlutWindow::runTimer(int _val) {
+void Window::runTimer(int _val) {
   current()->simTimer(_val);
 }
 
-bool GlutWindow::screenshot() {
+bool Window::screenshot() {
   static int count = 0;
   const char directory[8] = "frames";
   const char fileBase[8] = "Capture";
@@ -218,7 +219,7 @@ bool GlutWindow::screenshot() {
   }
 }
 
-inline GlutWindow* GlutWindow::current() {
+inline Window* Window::current() {
   int id = glutGetWindow();
   for (unsigned int i = 0; i < mWinIDs.size(); i++) {
     if (mWinIDs.at(i) == id) {
@@ -229,21 +230,22 @@ inline GlutWindow* GlutWindow::current() {
   exit(0);
 }
 
-void GlutWindow::keyboard(unsigned char /*_key*/, int /*_x*/, int /*_y*/) {
+void Window::keyboard(unsigned char /*_key*/, int /*_x*/, int /*_y*/) {
   // TODO(JS): Is 2d point information necessary for keyboard event?
 }
 
-void GlutWindow::specKey(int /*_key*/, int /*_x*/, int /*_y*/) {
+void Window::specKey(int /*_key*/, int /*_x*/, int /*_y*/) {
 }
 
-void GlutWindow::click(int /*_button*/, int /*_state*/, int /*_x*/, int /*_y*/) {
+void Window::click(int /*_button*/, int /*_state*/, int /*_x*/, int /*_y*/) {
 }
 
-void GlutWindow::drag(int /*_x*/, int /*_y*/) {
+void Window::drag(int /*_x*/, int /*_y*/) {
 }
 
-void GlutWindow::move(int /*_x*/, int /*_y*/) {
+void Window::move(int /*_x*/, int /*_y*/) {
 }
 
+}  // namespace glut
 }  // namespace gui
 }  // namespace dart
