+++ dart/gui/glut/GraphWindow.cpp
@@ -36,20 +36,97 @@
  *   POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef DART_GUI_SOFTSIMWINDOW_HPP_
-#define DART_GUI_SOFTSIMWINDOW_HPP_
+#include "dart/gui/glut/GraphWindow.hpp"
 
-#warning "This file is deprecated in DART 6.1. "\
-         "Please use dart/gui/glut/SoftSimWindow.hpp instead."
+#include <cstdio>
+#include <iostream>
+#include <string>
 
-#include "dart/gui/glut/SoftSimWindow.hpp"
+#include "dart/gui/glut/GLUTFuncs.hpp"
+#include "dart/gui/glut/LoadGlut.hpp"
 
 namespace dart {
 namespace gui {
+namespace glut {
 
-using SoftSimWindow = ::dart::gui::glut::SoftSimWindow;
+GraphWindow::GraphWindow()
+  : Win2D() {
+  mBackground[0] = 1.0;
+  mBackground[1] = 1.0;
+  mBackground[2] = 1.0;
+  mBackground[3] = 1.0;
+}
 
+GraphWindow::~GraphWindow() {
+}
+
+void GraphWindow::draw() {
+  mRI->setPenColor(Eigen::Vector3d(0.2, 0.8, 0.2));
+  glPointSize(2);
+  glMatrixMode(GL_MODELVIEW);
+
+  int nPoints = mData.size();
+
+  double upperBound = +1.0;
+  double lowerBound = -1.0;
+  if (nPoints > 0) {
+    upperBound = mData.maxCoeff();
+    lowerBound = mData.minCoeff();
+  }
+
+  for (int i = 0; i < nPoints; i++) {
+    glPushMatrix();
+    glLoadIdentity();
+    glBegin(GL_POINTS);
+    glVertex2f(i / (double)nPoints * mWinWidth - mWinWidth / 2.0, mWinHeight * (mData[i] - lowerBound) / (upperBound - lowerBound) - mWinHeight / 2.0);
+    glEnd();
+    glPopMatrix();
+  }    
+  glMatrixMode(GL_PROJECTION);
+
+  double xPos = 0.1;
+  while (xPos < 1.0) {
+    char buff[64];
+    int v = xPos * nPoints;
+#ifdef _WIN32
+    _snprintf(buff, sizeof(buff), "%d", v);
+#else
+    std::snprintf(buff, sizeof(buff), "%d", v);
+#endif
+    std::string frame(buff);
+    glColor3f(0.0, 0.0, 0.0);
+    drawStringOnScreen(xPos, 0.01f, frame, false);
+    xPos += 0.2;
+  }
+
+  double yPos = 0.1;
+  while (yPos < 1.0) {
+    char buff[64];
+    double v = yPos * (upperBound - lowerBound) + lowerBound;
+#ifdef _WIN32
+    _snprintf(buff, sizeof(buff), "%.2e", v);
+#else
+    std::snprintf(buff, sizeof(buff), "%.2e", v);
+#endif
+    std::string frame(buff);
+    glColor3f(0.0, 0.0, 0.0);
+    drawStringOnScreen(0.01f, yPos, frame, false);
+    yPos += 0.2;
+  }
+}
+
+void GraphWindow::keyboard(unsigned char _key, int _x, int _y) {
+  switch (_key) {
+    default:
+      Win2D::keyboard(_key, _x, _y);
+  }
+  glutPostRedisplay();
+}
+
+void GraphWindow::setData(Eigen::VectorXd _data) {
+  mData = _data;
+}
+
+}  // namespace glut
 }  // namespace gui
 }  // namespace dart
-
-#endif  // DART_GUI_SOFTSIMWINDOW_HPP_
