+++ dart/gui/glut/Win2D.cpp
@@ -30,16 +30,17 @@
  *   POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "dart/gui/Win2D.hpp"
+#include "dart/gui/glut/Win2D.hpp"
 
 #include <iostream>
 
-#include "dart/gui/LoadGlut.hpp"
+#include "dart/gui/glut/LoadGlut.hpp"
 
 namespace dart {
 namespace gui {
+namespace glut {
 
-Win2D::Win2D() : GlutWindow() {
+Win2D::Win2D() : glut::Window() {
   mTransX = 0;
   mTransY = 0;
   mTranslate = false;
@@ -154,5 +155,6 @@ void Win2D::initGL() {
   glEnable(GL_LINE_SMOOTH);
 }
 
+}  // namespace glut
 }  // namespace gui
 }  // namespace dart
