+++ dart/gui/glut/Win3D.cpp
@@ -30,18 +30,20 @@
  *   POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "dart/gui/Win3D.hpp"
+#include "dart/gui/glut/Win3D.hpp"
 
 #include <algorithm>
 
 #include "dart/math/Constants.hpp"
-#include "dart/gui/LoadGlut.hpp"
+#include "dart/gui/glut/LoadGlut.hpp"
 
 namespace dart {
 namespace gui {
+namespace glut {
 
+//==============================================================================
 Win3D::Win3D()
-  : GlutWindow(),
+  : Window(),
     mTrans(0.0, 0.0, 0.0),
     mEye(0.0, 0.0, 1.0),
     mUp(0.0, 1.0, 0.0),
@@ -49,17 +51,23 @@ Win3D::Win3D()
     mPersp(45.0),
     mRotate(false),
     mTranslate(false),
-    mZooming(false) {
+    mZooming(false)
+{
+  // Do nothing
 }
 
-void Win3D::initWindow(int _w, int _h, const char* _name) {
-  GlutWindow::initWindow(_w, _h, _name);
+//==============================================================================
+void Win3D::initWindow(int _w, int _h, const char* _name)
+{
+  Window::initWindow(_w, _h, _name);
 
   int smaller = _w < _h ? _w : _h;
   mTrackBall.setTrackball(Eigen::Vector2d(_w*0.5, _h*0.5), smaller/2.5);
 }
 
-void Win3D::resize(int _w, int _h) {
+//==============================================================================
+void Win3D::resize(int _w, int _h)
+{
   mWinWidth = _w;
   mWinHeight = _h;
 
@@ -79,7 +87,9 @@ void Win3D::resize(int _w, int _h) {
   glutPostRedisplay();
 }
 
-void Win3D::keyboard(unsigned char _key, int /*_x*/, int /*_y*/) {
+//==============================================================================
+void Win3D::keyboard(unsigned char _key, int /*_x*/, int /*_y*/)
+{
   switch (_key) {
     case ',':  // slow down
       mDisplayTimeout +=2;
@@ -109,7 +119,9 @@ void Win3D::keyboard(unsigned char _key, int /*_x*/, int /*_y*/) {
   // printf("ascii key: %lu\n", key);
 }
 
-void Win3D::click(int _button, int _state, int _x, int _y) {
+//==============================================================================
+void Win3D::click(int _button, int _state, int _x, int _y)
+{
   mMouseDown = !mMouseDown;
   int mask = glutGetModifiers();
   if (mMouseDown) {
@@ -141,7 +153,9 @@ void Win3D::click(int _button, int _state, int _x, int _y) {
   glutPostRedisplay();
 }
 
-void Win3D::drag(int _x, int _y) {
+//==============================================================================
+void Win3D::drag(int _x, int _y)
+{
   double deltaX = _x - mMouseX;
   double deltaY = _y - mMouseY;
 
@@ -162,7 +176,9 @@ void Win3D::drag(int _x, int _y) {
   glutPostRedisplay();
 }
 
-void Win3D::render() {
+//==============================================================================
+void Win3D::render()
+{
   glMatrixMode(GL_PROJECTION);
   glLoadIdentity();
   gluPerspective(mPersp,
@@ -220,7 +236,9 @@ void Win3D::render() {
     screenshot();
 }
 
-void Win3D::initGL() {
+//==============================================================================
+void Win3D::initGL()
+{
   glClearColor(mBackground[0], mBackground[1], mBackground[2], mBackground[3]);
   glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
   glEnable(GL_BLEND);
@@ -230,7 +248,9 @@ void Win3D::initGL() {
   glPolygonMode(GL_FRONT, GL_FILL);
 }
 
-void Win3D::initLights() {
+//==============================================================================
+void Win3D::initLights()
+{
   static float ambient[]             = {0.2, 0.2, 0.2, 1.0};
   static float diffuse[]             = {0.6, 0.6, 0.6, 1.0};
   static float front_mat_shininess[] = {60.0};
@@ -308,5 +328,6 @@ void accPerspective(GLdouble fovy, GLdouble aspect,
              pixdx, pixdy, eyedx, eyedy, focus);
 }
 
+}  // namespace glut
 }  // namespace gui
 }  // namespace dart
