+++ dart/gui/glut/glut.hpp.in
@@ -1,0 +1,3 @@
+// Automatically generated file by cmake
+
+${gui_glut_headers}
